+++ OneBlink.SDK/models/FormEvent.cs
@@ -1,0 +1,14 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormEvent
+    {
+        public string type
+        {
+            get; set;
+        }
+        public FormSubmissionEventConfiguration configuration
+        {
+            get; set;
+        }
+    }
+}